Fix the loops so they work. Ugh.
Somehow I thought the delay slot held the operation that affected the
jump. Of course, that makes no sense since the delay slot is the
pipelined instruction that has already been fetched meaning the jump has
already modified the PC for pre-fecth before it executes. This is what
happens when you only code after the kid goes down. Grr.