Stably works again.
Reenabled long fat filenames. (doh)

Move the song seleciton out of the button scan to avoid multiple
calls to start/stop when a button is pushed.

Remove the call to time(nullptr) in wake_incr() which avoids an
internal lock that would otherwise cause a deadlock if the interrupt
handler executes while the audio playback is occurring. (double doh).

There is still a race when you hit the button too fast, but this
at least works semi stably.